requirements.txt
hugchat==0.3.1
joblib==1.3.2
pandas==2.0.3
Requests==2.31.0
streamlit==1.27.2
scikit-learn==1.2.2
numpy==1.26.4
Pillow==9.4.0
tensorflow==2.11.0
opencv-python-headless==4.5.4.60
cryptography==38.0.4